DONE: - Refactored code to their respective folders/files. - Can now get a list of directions for plotting - Can now print mix of walking/bus route
